speed-up - only resolve path if needed
We need to resolve paths in case of symbolic links,
otherwise the mv intended to atomically replace the existing
file at the end of write_file will fail.
However, profiling reveals this to be expensive, so replace it
with a (cheaper) call to determine if the file is a symlink and
if so, resolve the path.